포인트 충전 로직 구현 테스트 로직 구현

+++ src/main/java/com/example/demo/point/PointRepository.java
@@ -13,5 +13,5 @@ public interface PointRepository extends JpaRepository<Point, Long> {
 
     @Transactional
     @Query("select sum(p.pointAmount) from Point p where p.user = ?1")
-    Long amountSum(User user);
+    Integer amountSum(User user);
 }
